+++ Spring-boot-demo/src/main/java/com/spring/project/demo/repositories/SpeakerRepository.java
@@ -3,5 +3,5 @@ package com.spring.project.demo.repositories;
 import com.spring.project.demo.models.Speakers;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpeakerRepository extends JpaRepository<Speakers, Long> {
+public interface SpeakerRepository extends JpaRepository<Speakers, Integer> {
 }
